The second argument type for marshalled actions is supposed to be size_t.
Change the LCO function definitions to "size_t" from an "int".

+++ examples/fibonacci.c
@@ -67,7 +67,7 @@ static int _fib_action(int *args, size_t size) {
     &fns[1]
   };
 
-  int sizes[] = {
+  size_t sizes[] = {
     sizeof(int),
     sizeof(int)
   };
